+++ src/main/java/com/buuz135/industrial/tile/mob/MobRelocatorTile.java
@@ -65,8 +65,8 @@ public class MobRelocatorTile extends WorkingAreaElectricMachine {
         this.getWorld().getEntitiesWithinAABB(EntityXPOrb.class, getWorkingArea().expand(2, 2, 2)).stream().filter(entityXPOrb -> !entityXPOrb.isDead).forEach(entityXPOrb -> {
             if (this.outExp.fill(new FluidStack(FluidsRegistry.ESSENCE, (int) (entityXPOrb.getXpValue() * 20 * ((MobRelocatorBlock) this.getBlockType()).getEssenceMultiplier())), false) > 0) {
                 this.outExp.fill(new FluidStack(FluidsRegistry.ESSENCE, (int) (entityXPOrb.getXpValue() * 20 * ((MobRelocatorBlock) this.getBlockType()).getEssenceMultiplier())), true);
-                entityXPOrb.setDead();
             }
+            entityXPOrb.setDead();
         });
     }