+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.5.10
 + Updated to Forge 1.12.2-14.23.1.2554
++ Fixed [#99](https://github.com/Buuz135/Industrial-Foregoing/issues/98)
 
 # 1.5.9
 + Fixed [#97](https://github.com/Buuz135/Industrial-Foregoing/issues/97)