+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Updated to Forge 1.12.2-14.23.1.2554
 + Fixed [#99](https://github.com/Buuz135/Industrial-Foregoing/issues/98)
 + The Mob Crusher now picks up the XP Orbs even if the tank is full
++ The Resourceful Furnace now doesn't allow to input items in the output slot
 
 # 1.5.9
 + Fixed [#97](https://github.com/Buuz135/Industrial-Foregoing/issues/97)